requirements.txt
django==3.2.1
pymysql==1.0.2
requests==2.27.1
djangorestframework==3.14.0
python-jenkins==1.8.0
cron-descriptor==1.4.0
apscheduler==3.10.1